ProtoUtils: duplicate -> rewind rather than save position.
Sometimes we see corrupt tokens being deserialized on application masters. I am not
sure what is causing this, but suspect a concurrency issue during job submission.
If this diff works it is a pretty weak "fix", although I haven't followed the
hadoop code enough to confirm or deny a concurrency issue. So we'll see what happens...