Revert commits ec96a04669d9a56a34650515f73aa4352d89cd03 and 19bc8450e3d2b3ceea6cd70be5006740a893996c.
These introduced debugging messages aimed at tracking down a concurrency issue in job submission. Turns
out the culprit is the JobConf(JobConf) constructor, which causes the created JobConf to share Credentials
with the parent JobConf. During job submission, the Credentials can be mutated, and since they are not
a thread-safe object this can cause it to get into an inconsistent state. We observed that the secret
token map had size() == 2 but its entrySet iterator only returned one element, which in turn caused
the serialized form of the Credentials to be corrupt, leading to a short read on deserialization.